requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pypdf>=4.0.0
python-docx>=0.8.11
python-dotenv>=1.0.0